feat: merge the 2 utils files and refactor
